# apps/server/internal/api/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Robot Center 관제 서버 API입니다. Robot, Operator, Recorder, System 역할별 API를 제공합니다.",
        "title": "Robot Center 관제 API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/operator/missions": {
            "get": {
                "description": "관제 서버에 등록된 임무 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "임무 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "로봇을 배정한 임무를 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "임무 생성",
                "parameters": [
                    {
                        "description": "임무 생성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.CreateMissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionEnvelopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/missions/{missionCode}/end": {
            "post": {
                "description": "active 상태 임무를 ended 상태로 전환하고 SFU room을 닫습니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "임무 종료",
                "parameters": [
                    {
                        "type": "string",
                        "description": "임무 코드",
                        "name": "missionCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionEnvelopeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/operator/missions/{missionCode}/start": {
            "post": {
                "description": "ready 상태 임무를 active 상태로 전환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "임무 시작",
                "parameters": [
                    {
                        "type": "string",
                        "description": "임무 코드",
                        "name": "missionCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionEnvelopeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "robot-center_apps_server_internal_api_dto.CreateMissionRequest": {
            "type": "object",
            "properties": {
                "missionType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "robotCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "siteNote": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.MissionEnvelopeResponse": {
            "type": "object",
            "properties": {
                "mission": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionResponse"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.MissionResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "endedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "missionCode": {
                    "type": "string"
                },
                "missionType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "robotCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "siteNote": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.MissionsResponse": {
            "type": "object",
            "properties": {
                "missions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.MissionResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "RobotBearerAuth": {
            "description": "Robot API에서 사용하는 Bearer robotToken입니다.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}